handles each camps without login

+++ XploreProject/Search CampSite/SearchCampVC.swift
@@ -473,27 +473,39 @@ class SearchCampVC: UIViewController, filterValuesDelegate {
 
     //MARK:- Button Action
     @IBAction func tapProfileBtn(_ sender: UIButton) {
-        let vc = self.storyboard?.instantiateViewController(withIdentifier: "MyProfileVC") as! MyProfileVC
-        self.navigationController?.pushViewController(vc, animated: true)
-        
+        if DataManager.isUserLoggedIn! {
+            let vc = self.storyboard?.instantiateViewController(withIdentifier: "MyProfileVC") as! MyProfileVC
+            self.navigationController?.pushViewController(vc, animated: true)
+        } else {
+            self.loginAlertFunc(vc: "profile")
+        }
     }
     
     @IBAction func tapNearByUserBtn(_ sender: UIButton) {
-        let vc = self.storyboard?.instantiateViewController(withIdentifier: "NearByUsersVC") as! NearByUsersVC
-        self.navigationController?.pushViewController(vc, animated: true)
-        
+        if DataManager.isUserLoggedIn! {
+            let vc = self.storyboard?.instantiateViewController(withIdentifier: "NearByUsersVC") as! NearByUsersVC
+            self.navigationController?.pushViewController(vc, animated: true)
+        } else {
+            self.loginAlertFunc(vc: "nearByUser")
+        }
     }
     
     @IBAction func tapAddCampsiteBtn(_ sender: UIButton) {
-        let vc = self.storyboard?.instantiateViewController(withIdentifier: "AddNewCampsiteVc") as! AddNewCampsiteVc
-        self.navigationController?.pushViewController(vc, animated: true)
-        
+        if DataManager.isUserLoggedIn! {
+            let vc = self.storyboard?.instantiateViewController(withIdentifier: "AddNewCampsiteVc") as! AddNewCampsiteVc
+            self.navigationController?.pushViewController(vc, animated: true)
+        } else {
+            self.loginAlertFunc(vc: "addCamps")
+        }
     }
     
     @IBAction func tapNotificationBtn(_ sender: UIButton) {
-        let vc = self.storyboard?.instantiateViewController(withIdentifier: "NotificationVc") as! NotificationVc
-        self.navigationController?.pushViewController(vc, animated: true)
-        
+        if DataManager.isUserLoggedIn! {
+            let vc = self.storyboard?.instantiateViewController(withIdentifier: "NotificationVc") as! NotificationVc
+            self.navigationController?.pushViewController(vc, animated: true)
+        } else {
+            self.loginAlertFunc(vc: "fromNoti")
+        }
     }
     
     @IBAction func backAction(_ sender: Any) {
@@ -595,23 +607,27 @@ extension SearchCampVC {
     }
     
     @objc func favoutiteAction(sender: UIButton) {
-        self.campIndex = sender.tag
-        if String(describing: ((self.searchDataArr.object(at: self.campIndex) as! NSDictionary).value(forKey: "isFav"))!) == "0" {
-            self.markAsFavBtn.setTitle("Mark as favourite", for: .normal)
+        if DataManager.isUserLoggedIn! {
+            self.campIndex = sender.tag
+            if String(describing: ((self.searchDataArr.object(at: self.campIndex) as! NSDictionary).value(forKey: "isFav"))!) == "0" {
+                self.markAsFavBtn.setTitle("Mark as favourite", for: .normal)
+                
+            } else {
+                self.markAsFavBtn.setTitle("Delete from favourite", for: .normal)
+                
+            }
             
+            self.campId = Int(String(describing: ((self.searchDataArr.object(at: sender.tag) as! NSDictionary).value(forKey: "campId"))!))!
+            
+            self.favMarkbottomConstraint.constant = 150
+            
+            self.overlayview.tag = sender.tag
+            self.overlayview.isHidden = false
+            self.view.layoutIfNeeded()
         } else {
-            self.markAsFavBtn.setTitle("Delete from favourite", for: .normal)
-            
+            self.loginAlertFunc(vc: "markFav")
+            Singleton.sharedInstance.favIndex = sender.tag
         }
-        
-        self.campId = Int(String(describing: ((self.searchDataArr.object(at: sender.tag) as! NSDictionary).value(forKey: "campId"))!))!
-        
-        self.favMarkbottomConstraint.constant = 150
-        
-        self.overlayview.tag = sender.tag
-        self.overlayview.isHidden = false
-        self.view.layoutIfNeeded()
-        
     }
 }
 
@@ -713,17 +729,22 @@ extension SearchCampVC :UICollectionViewDataSource ,UICollectionViewDelegate {
     }
     
     @objc func tapSearchProfilePicBtn(sender: UIButton) {
-        let indexVal: NSDictionary = (self.searchDataArr.object(at: sender.tag) as! NSDictionary)
-        
-        if let campAuth = indexVal.value(forKey: "campAuthor") as? String {
-            if String(describing: (DataManager.userId)) == campAuth {
-                let vc = self.storyboard?.instantiateViewController(withIdentifier: "MyProfileVC") as! MyProfileVC
-                self.navigationController?.pushViewController(vc, animated: true)
-            } else {
-                let vc = self.storyboard?.instantiateViewController(withIdentifier: "UserProfileVC") as! UserProfileVC
-                vc.userInfoDict = indexVal
-                self.navigationController?.pushViewController(vc, animated: true)
-                
+        if DataManager.isUserLoggedIn! == false {
+            self.loginAlertFunc(vc: "viewProfile")
+            
+        } else {
+            let indexVal: NSDictionary = (self.searchDataArr.object(at: sender.tag) as! NSDictionary)
+            
+            if let campAuth = indexVal.value(forKey: "campAuthor") as? String {
+                if String(describing: (DataManager.userId)) == campAuth {
+                    let vc = self.storyboard?.instantiateViewController(withIdentifier: "MyProfileVC") as! MyProfileVC
+                    self.navigationController?.pushViewController(vc, animated: true)
+                } else {
+                    let vc = self.storyboard?.instantiateViewController(withIdentifier: "UserProfileVC") as! UserProfileVC
+                    vc.userInfoDict = indexVal
+                    self.navigationController?.pushViewController(vc, animated: true)
+                    
+                }
             }
         }
     }
@@ -779,5 +800,44 @@ extension SearchCampVC: GMSAutocompleteViewControllerDelegate {
         self.dataContainingView.isHidden = true
         dismiss(animated: true, completion: nil)
         self.navigationController?.popViewController(animated: false)
+    }
+}
+
+//MARK:- login alert
+extension SearchCampVC {
+    func loginAlertFunc(vc: String) {
+        let alert = UIAlertController(title: appName, message: loginRequired, preferredStyle: .alert)
+        let yesBtn = UIAlertAction(title: Ok, style: .default, handler: { (UIAlertAction) in
+            alert.dismiss(animated: true, completion: nil)
+            let controller = self.storyboard?.instantiateViewController(withIdentifier: "LoginVc") as! LoginVc
+            if vc == "profile" {
+                Singleton.sharedInstance.loginComeFrom = fromProfile
+                
+            } else if vc == "nearByUser" {
+                Singleton.sharedInstance.loginComeFrom = fromNearByuser
+               
+            } else if vc == "addCamps" {
+                Singleton.sharedInstance.loginComeFrom = fromAddCamps
+                
+            } else if vc == "fromNoti" {
+                Singleton.sharedInstance.loginComeFrom = fromNoti
+                
+            } else if vc == "fromNoti" {
+                Singleton.sharedInstance.loginComeFrom = fromFavCamps
+                
+            } else if vc == "viewProfile" {
+                Singleton.sharedInstance.loginComeFrom = fromViewProfile
+                
+            }
+            self.navigationController?.pushViewController(controller, animated: false)
+        })
+        
+        let noBtn = UIAlertAction(title: cancel, style: .default, handler: { (UIAlertAction) in
+            alert.dismiss(animated: true, completion: nil)
+        })
+        alert.addAction(yesBtn)
+        alert.addAction(noBtn)
+        present(alert, animated: true, completion: nil)
+        
     }
 }
